requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
boto3>=1.26.0
pillow>=10.0.0,<11.0.0  # Try to get a compatible version
matplotlib>=3.7.0
numpy>=1.24.0
python-multipart>=0.0.6
pydantic>=2.0.0,<3.0.0
python-dotenv>=0.21.0